% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/azureApi.R
\name{azGetBlob}
\alias{azGetBlob}
\title{The Get Blob operation reads or downloads a blob from the system, including its metadata and properties. You can also call Get Blob to read a snapshot.}
\usage{
azGetBlob(account)
}
\description{
The Get Blob operation reads or downloads a blob from the system, including its metadata and properties. You can also call Get Blob to read a snapshot.
}
\references{
https://msdn.microsoft.com/en-us/library/azure/dd179440.aspx
}

