% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/azureApi.R
\name{azPutBlob}
\alias{azPutBlob}
\title{The Put Blob operation creates a new block, page, or append blob, or updates the content of an existing block blob.}
\usage{
azPutBlob(account, blob)
}
\description{
The Put Blob operation creates a new block, page, or append blob, or updates the content of an existing block blob.
}
\references{
https://msdn.microsoft.com/en-us/library/azure/dd179451.aspx
}

