% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AzureBlob-package.R
\docType{package}
\name{AzureBlob-package}
\alias{AzureBlob}
\alias{AzureBlob-package}
\title{Interface to Azure Blob Storage}
\description{
Allows you to work with Azure ML Studio datasets and experiments directly from R.
}
\section{Summary of functions}{


1. Create a link to an Azure ML workspace

\itemize{
\item Link to workspace: \code{\link{workspace}}
}

2. Datasets

\itemize{
   \item Get datasets: \code{\link{datasets}}
}

3. Experiments

\itemize{
   \item Get experiments: \code{\link{experiments}}
}
}
\keyword{package}

